refactor: 스웨거 인터페이스 적용
