+++ src/main/java/oncog/cogroom/domain/auth/controller/AuthController.java
@@ -6,6 +6,7 @@ import jakarta.servlet.http.HttpServletResponse;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import oncog.cogroom.domain.auth.docs.AuthControllerDocs;
 import oncog.cogroom.domain.auth.dto.request.AuthRequestDTO;
 import oncog.cogroom.domain.auth.exception.AuthErrorCode;
 import oncog.cogroom.domain.auth.service.AuthServiceRouter;
@@ -31,17 +32,13 @@ import static oncog.cogroom.domain.auth.dto.response.AuthResponseDTO.SignupRespo
 @Slf4j
 @RequiredArgsConstructor
 @RequestMapping("/api/v1/auth")
-public class AuthController {
+public class AuthController implements AuthControllerDocs {
 
     private final AuthServiceRouter router;
     private final EmailService emailService;
     private final CookieUtil cookieUtil;
 
     @PostMapping("/login")
-    @ApiErrorCodeExample(
-            value = AuthErrorCode.class,
-            include = {"KAKAO_AUTH_FAILED", "KAKAO_INVALID_AUTHORIZATION_CODE"})
-    @Operation(summary = "소셜/로컬 통합 로그인", description = "소셜/로컬 통합 로그인 로직을 처리합니다. \n 응답 코드에 따른 자세한 결과는 Notion 명세서를 참고 부탁드립니다.")
     public ResponseEntity<ApiResponse<LoginResponseDTO>> socialLogin(@RequestBody @Valid LoginRequestDTO request, HttpServletResponse response) {
         LoginResponseDTO result = router.login(request);
 
@@ -56,10 +53,6 @@ public class AuthController {
 
 
     @PostMapping("/signup")
-    @ApiErrorCodeExamples(
-            value = {AuthErrorCode.class, ApiErrorCode.class},
-            include = {"EMPTY_FILED", "INVALID_EMAIL_FOTMAT"})
-    @Operation(summary = "소셜/로컬 통합 회원가입", description = "소셜/로컬 통합 회원가입 로직을 처리합니다. \n 응답 코드에 따른 자세한 결과는 Notion 명세서를 참고 부탁드립니다.")
     public ResponseEntity<ApiResponse<SignupResponseDTO>> socialSignup(@RequestBody @Valid SignupRequestDTO request, HttpServletResponse response) {
         SignupResponseDTO result = router.signup(request);
 
@@ -73,13 +66,11 @@ public class AuthController {
     }
 
     @PostMapping("/email-verification")
-    @ApiErrorCodeExamples(
-            value = {AuthErrorCode.class, ApiErrorCode.class},
-            include = {"EMPTY_FILED", "INVALID_EMAIL_FOTMAT", "ALREADY_EXIST_EMAIL"})
-    @Operation(summary = "인증 이메일 전송", description = "인증용 링크가 포함된 이메일을 전송합니다. \n 응답 코드에 따른 자세한 결과는 Notion 명세서를 참고 부탁드립니다.")
     public ResponseEntity<ApiResponse<String>> sendEmail(@RequestBody @Valid AuthRequestDTO.EmailRequestDTO request) throws MessagingException, IOException {
+        // 이메일 중복 검사
         emailService.existEmail(request.getEmail());
 
+        // 중복 검사 이후 이메일 전송
         emailService.sendEmail(request);
 
         return ResponseEntity.ok(ApiResponse.of(ApiSuccessCode.SUCCESS));
@@ -87,10 +78,6 @@ public class AuthController {
     }
 
     @GetMapping("/check-verification")
-    @ApiErrorCodeExamples(
-            value = {AuthErrorCode.class, ApiErrorCode.class},
-            include = {"EMPTY_FILED","EXPIRED_LINK"})
-    @Operation(summary = "이메일 인증", description = "링크가 클릭되었을 때 이메일을 인증합니다. \n 응답 코드에 따른 자세한 결과는 Notion 명세서를 참고 부탁드립니다.")
     public ResponseEntity<ApiResponse<Void>> verifyEmail(@RequestParam String userEmail,
                                                          @RequestParam String verificationCode) {
         emailService.verifyCode(userEmail,verificationCode);
@@ -99,11 +86,8 @@ public class AuthController {
     }
 
     @PostMapping("/email/status")
-    @Operation(summary = "이메일 인증 여부 반환", description = "이메일의 인증이 완료되었는지 여부를 반환합니다. \n 응답 코드에 따른 자세한 결과는 Notion 명세서를 참고 부탁드립니다.")
     public ResponseEntity<ApiResponse<Boolean>> checkEmailVerificationStatus(@RequestBody @Valid AuthRequestDTO.EmailRequestDTO request) {
-        boolean result = emailService.verifiedEmail(request);
-
-        return ResponseEntity.ok(ApiResponse.of(ApiSuccessCode.SUCCESS, result));
+        return ResponseEntity.ok(ApiResponse.of(ApiSuccessCode.SUCCESS, emailService.verifiedEmail(request)));
 
     }
 }
